half functionality from chromadb
