+++ app/api/chroma/fetch-cloud-data/route.ts
@@ -1,4 +1,4 @@
-// For adding to collection
+// For creating a collection
 
 import client from '@/lib/chroma';
 import { NextRequest, NextResponse } from 'next/server';
@@ -8,24 +8,20 @@ export const dynamic = "force-dynamic";
 export async function POST(req: NextRequest) {
   try {
     const body = await req.json();
-    const documents = body.documents;
-    const filename = body.filename;
 
-    const ids = Array.from({ length: documents.length }, (_, i) => (i + 1).toString());
-    const metadatas = Array.from({ length: documents.length }, () => ({ file_name: `${filename}` }));
+    const cloud_name = body.cloud_name;
 
     const collection = await client.getOrCreateCollection({
       name: 'myCollection',
       metadata: { 'description': `Collection for user: ${"myCollection"}` }
     });
 
-    await collection.add({
-      ids: ids,
-      documents: documents,
-      metadatas: metadatas
-    });
+    const resp = await collection.query({
+      queryTexts: [`Cloud: ${cloud_name}`],
+      where: { cloud: cloud_name }
+    })
 
-    return NextResponse.json({ response: "success" });
+    return NextResponse.json({ response: resp });
   } catch(error: any) {
   
     console.error("Error in API:", error);
